requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.28.0
python-multipart==0.0.6
pydantic==2.4.2
typing-extensions>=4.8.0

# OpenAI API
openai==1.3.0

# Image processing
pillow==10.1.0
requests==2.31.0

# Testing
pytest==7.4.3
httpx==0.25.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
psutil==5.9.5